+++ src/Form/DependentContentRevisionRevertForm.php
@@ -5,22 +5,18 @@ namespace Drupal\dependent_content\Form;
 use Drupal\Component\Datetime\TimeInterface;
 use Drupal\Core\Datetime\DateFormatterInterface;
 use Drupal\Core\Entity\EntityStorageInterface;
+use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Form\ConfirmFormBase;
 use Drupal\Core\Form\FormStateInterface;
+use Drupal\Core\Language\LanguageManagerInterface;
 use Drupal\Core\Url;
+use Drupal\dependent_content\Entity\DependentContentInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Provides a form for reverting an entity revision.
  */
 class DependentContentRevisionRevertForm extends ConfirmFormBase {
-
-  /**
-   * The entity revision
-   *
-   * @var \Drupal\Core\Entity\EntityInterface|\Drupal\Core\Entity\RevisionLogInterface|\Drupal\Core\Entity\EntityChangedInterface $revision
-   */
-  protected $revision;
 
   /**
    * The entity storage.
@@ -44,35 +40,48 @@ class DependentContentRevisionRevertForm extends ConfirmFormBase {
   protected $time;
 
   /**
+   * The language manager.
+   *
+   * @var \Drupal\Core\Language\LanguageManagerInterface
+   */
+  protected $languageManager;
+
+  /**
    * DependentContentRevisionRevertForm constructor.
    *
-   * @param \Drupal\Core\Entity\EntityStorageInterface $storage
+   * @param EntityTypeManagerInterface $manager
    *   The entity storage.
    * @param \Drupal\Core\Datetime\DateFormatterInterface $date_formatter
    *   The date formatter service.
    * @param \Drupal\Component\Datetime\TimeInterface $time
    *   The time service.
+   * @param \Drupal\Core\Language\LanguageManagerInterface $language_manager
+   *   The language manager.
    */
-  public function __construct(EntityStorageInterface $storage, DateFormatterInterface $date_formatter, TimeInterface $time) {
+  public function __construct(EntityTypeManagerInterface $manager, DateFormatterInterface $date_formatter, TimeInterface $time, LanguageManagerInterface $language_manager) {
 
-    $this->storage = $storage;
+    $this->storage = $manager->getStorage('dependent_content');
     $this->dateFormatter = $date_formatter;
     $this->time = $time;
+    $this->languageManager = $language_manager;
   }
 
   public static function create(ContainerInterface $container) {
 
-    /** @var \Drupal\Core\Entity\EntityStorageInterface $storage */
-    $storage = $container->get('entity_type.manager')->getStorage('dependent_content');
+    /** @var EntityTypeManagerInterface $manager */
+    $manager = $container->get('entity_type.manager');
     /** @var \Drupal\Core\Datetime\DateFormatterInterface $date_formatter */
     $date_formatter = $container->get('date.formatter');
     /** @var \Drupal\Component\Datetime\TimeInterface $time */
     $time = $container->get('datetime.time');
+    /** @var LanguageManagerInterface $language_manager */
+    $language_manager = $container->get('language_manager');
 
     return new static(
-      $storage,
+      $manager,
       $date_formatter,
-      $time
+      $time,
+      $language_manager
     );
   }
 
@@ -84,8 +93,12 @@ class DependentContentRevisionRevertForm extends ConfirmFormBase {
    */
   public function getQuestion() {
 
+    $dependent_content_revision = $this->getRequest()->get('dependent_content_revision');
+    /** @var DependentContentInterface $revision */
+    $revision = $this->storage->loadRevision($dependent_content_revision);
+
     return $this->t('Are you sure you want to revert to the new revision from %date', array(
-      '%date' => $this->dateFormatter->format($this->revision->getRevisionCreationTime(), 'short')
+      '%date' => $this->dateFormatter->format($revision->getRevisionCreationTime(), 'short')
     ));
   }
 
@@ -109,7 +122,7 @@ class DependentContentRevisionRevertForm extends ConfirmFormBase {
   public function getCancelUrl() {
 
     return new Url('entity.dependent_content_revision.history', array(
-      'dependent_content' => $this->revision->id()
+      'dependent_content' => $this->getRequest()->get('dependent_content')
     ));
   }
 
@@ -139,7 +152,15 @@ class DependentContentRevisionRevertForm extends ConfirmFormBase {
    */
   public function buildForm(array $form, FormStateInterface $form_state, $dependent_content_revision = NULL) {
 
-    $this->revision = $this->storage->loadRevision($dependent_content_revision);
+    $langcode = $this->languageManager->getCurrentLanguage()->getId();
+    /** @var DependentContentInterface $revision */
+    $revision = $this->storage->loadRevision($dependent_content_revision);
+
+    if ($revision->hasTranslation($langcode)) {
+      $revision = $revision->getTranslation($langcode);
+    }
+
+    $form_state->set('revision', $revision);
 
     return parent::buildForm($form, $form_state);
   }
@@ -154,33 +175,37 @@ class DependentContentRevisionRevertForm extends ConfirmFormBase {
    */
   public function submitForm(array &$form, FormStateInterface $form_state) {
 
-    $original_timestamp = $this->revision->getRevisionCreationTime();
+    /** @var DependentContentInterface $revision */
+    $revision = $form_state->get('revision');
+
+    $original_timestamp = $revision->getRevisionCreationTime();
     $log_message = $this->t('Copy of the revision from %date', array(
       '%date' => $this->dateFormatter->format($original_timestamp, 'short')
     ));
 
-    $this->revision->setNewRevision();
-    $this->revision->isDefaultRevision(TRUE);
-    $this->revision->setRevisionLogMessage($log_message);
-    $this->revision->setRevisionCreationTime($this->time->getRequestTime());
-    $this->revision->setChangedTime($this->time->getRequestTime());
+    $revision->setNewRevision();
+    $revision->isDefaultRevision(TRUE);
 
-    $this->revision->save();
+    $revision->setRevisionLogMessage($log_message);
+    $revision->setRevisionCreationTime($this->time->getRequestTime());
+    $revision->setChangedTime($this->time->getRequestTime());
+
+    $revision->save();
 
     $this->logger('dependent content')->notice('%type: reverted %title revision %revision.', array(
-      '%type' => $this->revision->bundle(),
-      '%title' => $this->revision->label(),
-      '%revision' => $this->revision->getRevisionId()
+      '%type' => $revision->bundle(),
+      '%title' => $revision->label(),
+      '%revision' => $revision->getRevisionId()
     ));
 
     drupal_set_message($this->t('%type %title has been reverted to the revision from %date.', array(
-      '%type' => $this->revision->getEntityType()->getLabel(),
-      '%title' => $this->revision->label(),
+      '%type' => $revision->getEntityType()->getLabel(),
+      '%title' => $revision->label(),
       '%date' => $this->dateFormatter->format($original_timestamp)
     )));
 
     $form_state->setRedirect('entity.dependent_content_revision.history', array(
-      'dependent_content' => $this->revision->id()
+      'dependent_content' => $this->getRequest()->get('dependent_content')
     ));
   }
 }